refactor(api): migrate group repository to its bounded context
