+++ api/src/prescription/campaign/domain/models/Group.js
