+++ api/src/prescription/campaign/infrastructure/repositories/group-repository.js
@@ -1,4 +1,4 @@
-import { knex } from '../../../db/knex-database-connection.js';
+import { knex } from '../../../../../db/knex-database-connection.js';
 import { Group } from '../../domain/models/Group.js';
 
 async function findByCampaignId(campaignId) {
